test updates

+++ app/models/user.rb
@@ -349,16 +349,16 @@ class User < ActiveRecord::Base
     ActiveRecord::Base.connection.execute(
       <<-SQL
         SELECT
-          CASE WHEN (EXTRACT(month from created_at)) > 9
-               THEN extract(year from created_at)+1
-               ELSE extract(year from created_at)
+          CASE WHEN (EXTRACT(month from created_at))::int > 9
+               THEN extract(year from created_at)::int + 1
+               ELSE extract(year from created_at)::int
             END
         FROM readings
         WHERE user_id = #{id}
         GROUP BY
-          CASE WHEN (EXTRACT(month from created_at)) > 9
-               THEN extract(year from created_at)+1
-               ELSE extract(year from created_at)
+          CASE WHEN (EXTRACT(month from created_at))::int > 9
+               THEN extract(year from created_at)::int + 1
+               ELSE extract(year from created_at)::int
             END;
       SQL
     ).to_a.map { |r| r["date_part"].to_i }.map do |year|
